Infra.Automation-pipe_image/requirements.txt
urllib3==1.25.9
requests==2.24.0
PyYAML==5.3.1
kubernetes==12.0.1
openshift==0.12.1
python-jenkins==1.0.1
pytz>=2014.4
requests>=2.3.0
six>=1.10.0
requests-toolbelt
